+++ test/cli.test.js
@@ -2458,6 +2458,7 @@ describe('[Common]', () => {
       expect(chains.mainnet.native).toBe(undefined);
       expect(chains.bellecour.native).toBe(undefined);
       expect(chains.kovan.native).toBe(undefined);
+      expect(chains.goerli.native).toBe(undefined);
     }, 10000);
 
     test('mainnet is not native', async () => {
@@ -2474,6 +2475,17 @@ describe('[Common]', () => {
     test('kovan is not native', async () => {
       const raw = await execAsync(
         `${iexecPath} wallet show ${ADDRESS} --chain kovan --raw`,
+      );
+      const res = JSON.parse(raw);
+      expect(res.ok).toBe(true);
+      expect(res.balance).not.toBe(undefined);
+      expect(res.balance.nRLC).not.toBe(undefined);
+      expect(res.balance.ETH).not.toBe(undefined);
+    }, 10000);
+
+    test('goerli is not native', async () => {
+      const raw = await execAsync(
+        `${iexecPath} wallet show ${ADDRESS} --chain goerli --raw`,
       );
       const res = JSON.parse(raw);
       expect(res.ok).toBe(true);