+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- `gpu` tag is now supported
+
 ### Changed
 
 ### Removed